+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed a bug in coin selection where the wallet's mixed inputs were not considered for a `TxBodyContent` with zero inputs
 * Considering withdrawals and certificates in `mapTxScriptWtinesses`.
 * Function `substituteExecutionUnits` updated to trigger `TxBodyErrorAutoBalance` when witnesses are not found.
+* `MonadMockchain.sendTx` returns `Either` to allow the caller to handle submission failures
 
 ### Added